KIN-185 Added project structure and updated readme
The project structure based on code igniter has been added as well as gulp has been added for the compilation of the assets. 

Has been added the essentials to make use of velzon and modified the readme as part of the documentation, in the areas of project structure, project prerequisites, installation process, and commands for run.