KIN-386 Enhance interactive video experience and page scrolling UX
This update focuses on improving the video experience and scrolling UX for mobile and desktop users. For the video, we've tweaked settings so that the video starts playing when it is 80% visible on the screen, providing an interactive experience for users. If the video is not visible, it will pause, saving resources. We've removed the auto-mute feature but set a reasonable default volume level to balance user engagement and comfort.

For scrolling, we've implemented scroll-snap and adjusted scroll-padding and scroll-margin values to smooth out and optimize the scrolling behavior across various sections of the site. The changes influence how elements align when users navigate via scrolling especially, taking various device screen sizes into consideration.

These changes collectively aim to enhance user engagement and provide a smoother, more consistent browsing experience across devices.